Fix bug positioning on same cell two rooms
