--- Shootingat Resolved Project/Assets/Scripts/Dungeon Generation/DungeonGeneration.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ac07743724e78fc40bad17f884cf7d4e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
