+++ Shootingat Resolved Project/UserSettings/EditorUserSettings.asset
@@ -30,10 +30,10 @@ EditorUserSettings:
       value: 22424703114646680e0b0227036c7d1f05042b292926237e38271427fb
       flags: 0
     RecentlyUsedScenePath-8:
-      value: 22424703114646680e0b0227036c78111b122b292926237e38271427fb
+      value: 22424703114646680e0b0227036c78111b122b292926230f192c0e27c6f53a1fe7e933fbe62e3633374cfa320d2a18
       flags: 0
     RecentlyUsedScenePath-9:
-      value: 22424703114646680e0b0227036c78111b122b292926230f192c0e27c6f53a1fe7e933fbe62e3633374cfa320d2a18
+      value: 22424703114646680e0b0227036c78111b122b292926237e38271427fb
       flags: 0
     vcSharedLogLevel:
       value: 0d5e400f0650
